+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -12,6 +12,7 @@
 #include "Rule10p2Check.h"
 #include "Rule10p3Check.h"
 #include "Rule10p4Check.h"
+#include "Rule10p5Check.h"
 
 namespace clang::tidy {
 namespace misrac {
@@ -27,6 +28,8 @@ public:
         "misrac-rule-10p3");
     CheckFactories.registerCheck<Rule10p4Check>(
         "misrac-rule-10p4");
+    CheckFactories.registerCheck<Rule10p5Check>(
+        "misrac-rule-10p5");
   }
 };